Advertise conn pref: Diego cell IP vs Instance IP
Silk prefers `cf ssh` to Diego cell IP, and that's the default.

Third party network plugins will now have the option to advertise the
instance (container) IP address when they are placed on a separate
overlay network.

Set the `rep` BOSH job property
`advertise_preference_for_instance_address` to true to direct `cf ssh`
to use instance IP address.

This is a more fine-grained version of the `ssh-proxy` BOSH job
property, `connect_to_instance_address`, which, when set to `true`,
directs _all_ `cf ssh` connections to the instance's IP address, which
is the desired behavior when _all_ instances are on the overlay network.

Fine-grain is required for networks where not all instances are on the
overlay networks, for example, networks with a mixture of Linux and
Windows Diego cells.

By setting the preference for instance addresses on each `rep` job,
heterogenous deployments can specify the the best/accessible address on
the BOSH instance group level, allowing Linux cells to elect connecting to
instance addresses and vice versa.

[fixes #427]

[#166166339](https://www.pivotaltracker.com/story/show/166166339)

Signed-off-by: Brian Cunnie <bcunnie@pivotal.io>